+++ rotkehlchen/chain/evm/decoding/uniswap/v3/decoder.py
@@ -1,7 +1,6 @@
 import logging
 from typing import TYPE_CHECKING, Any
 
-from rotkehlchen.chain.decoding.constants import CPT_GAS
 from rotkehlchen.chain.decoding.types import (
     CounterpartyDetails,
     get_versioned_counterparty_label,
@@ -31,21 +30,18 @@ from rotkehlchen.chain.evm.decoding.uniswap.v3.constants import (
 )
 from rotkehlchen.chain.evm.decoding.uniswap.v3.utils import (
     decode_uniswap_v3_like_deposit_or_withdrawal,
+    decode_uniswap_v3_like_router_swap,
 )
-from rotkehlchen.chain.evm.structures import EvmTxReceiptLog, SwapData
-from rotkehlchen.constants import ZERO
 from rotkehlchen.errors.misc import RemoteError
-from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.logging import RotkehlchenLogsAdapter
-from rotkehlchen.utils.misc import ts_ms_to_sec
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
-    from rotkehlchen.assets.asset import Asset, EvmToken
+    from rotkehlchen.assets.asset import EvmToken
     from rotkehlchen.chain.evm.decoding.base import BaseEvmDecoderTools
     from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
-    from rotkehlchen.fval import FVal
+    from rotkehlchen.chain.evm.structures import EvmTxReceiptLog
     from rotkehlchen.history.events.structures.evm_event import EvmEvent
     from rotkehlchen.types import (
         ChecksumEvmAddress,
@@ -55,31 +51,6 @@ if TYPE_CHECKING:
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
-
-
-def _find_from_asset_and_amount(events: list[EvmEvent]) -> tuple[Asset, FVal] | None:
-    """
-    Searches for uniswap v2/v3 swaps, detects `from_asset` and sums up `from_amount`.
-    Works only with `from_asset` being the same for all swaps.
-    """
-    from_asset: Asset | None = None
-    from_amount = ZERO
-    for event in events:
-        if (
-            event.event_type == HistoryEventType.TRADE and
-            event.event_subtype == HistoryEventSubType.SPEND and
-            event.counterparty in {CPT_UNISWAP_V2, CPT_UNISWAP_V3}
-        ):
-            if from_asset is None:
-                from_asset = event.asset
-            elif from_asset != event.asset:  # We currently support only single `from_asset`.
-                return None  # unexpected event
-
-            from_amount += event.amount
-
-    if from_asset is None:
-        return None
-    return from_asset, from_amount
 
 
 class Uniswapv3CommonDecoder(EvmDecoderInterface):
@@ -99,40 +70,6 @@ class Uniswapv3CommonDecoder(EvmDecoderInterface):
         )
         self.routers_addresses = routers_addresses
         self.nft_manager = nft_manager
-
-    def _find_to_asset_and_amount(self, events: list[EvmEvent]) -> tuple[Asset, FVal] | None:
-        """
-        Searches for uniswap v2/v3 swaps, detects `to_asset` and sums up `to_amount`.
-        Works only with `to_asset` being the same for all swaps.
-        Also works with a special case where there is only one receive event at the end.
-        """
-        to_asset: Asset | None = None
-        to_amount = ZERO
-        for event in events:
-            if (
-                event.event_type == HistoryEventType.TRADE and
-                event.event_subtype == HistoryEventSubType.RECEIVE and
-                event.counterparty in {CPT_UNISWAP_V2, CPT_UNISWAP_V3}
-            ):
-                # Some swaps have many receive events. The structure is:
-                # spend1, receive1, spend2, receive2, ..., spendN, receiveN
-                # In this case they will be decoded as trades and we check them here.
-                if to_asset is None:
-                    to_asset = event.asset
-                elif to_asset != event.asset:  # We currently support only single `to_asset`.
-                    return None  # unexpected event
-                to_amount += event.amount
-            elif event.event_type == HistoryEventType.RECEIVE and event.asset != self.node_inquirer.native_token and to_asset is None:  # noqa: E501
-                # Some other swaps have only a single receive event. The structure is:
-                # spend1, spend2, ..., spendN, receive
-                # In this case the receive event won't be decoded as a trade and we check it here.
-                # to_asset should be None here since it should be the only receive event.
-                to_asset = event.asset
-                to_amount = event.amount
-
-        if to_asset is None:
-            return None
-        return to_asset, to_amount
 
     def _decode_deposits_and_withdrawals(self, context: DecoderContext) -> EvmDecodingOutput:
         if context.tx_log.topics[0] == INCREASE_LIQUIDITY_SIGNATURE:
@@ -214,67 +151,6 @@ class Uniswapv3CommonDecoder(EvmDecoderInterface):
             native_currency=self.node_inquirer.native_token,
         )
 
-    # --- Routers methods ---
-
-    def _decode_native_to_token_swap(
-            self,
-            decoded_events: list[EvmEvent],
-            send_native_event: EvmEvent,
-            receive_native_event: EvmEvent | None,
-    ) -> SwapData | None:
-        """
-        Decode a swap of native currency to a token. Such swap consists of 3 events:
-        1. Sending native currency to the router.
-        2. If there is a refund, receiving native currency from the router.
-        3. Receiving tokens from the router.
-        """
-        from_amount = send_native_event.amount
-        if receive_native_event is not None:
-            from_amount -= receive_native_event.amount  # a refund
-
-        to_data = self._find_to_asset_and_amount(decoded_events)
-        if to_data is None:
-            return None
-
-        return SwapData(
-            from_asset=self.node_inquirer.native_token,
-            from_amount=from_amount,
-            to_asset=to_data[0],
-            to_amount=to_data[1],
-        )
-
-    def _decode_token_to_native_swap(
-            self,
-            decoded_events: list[EvmEvent],
-            receive_native_event: EvmEvent,
-    ) -> SwapData | None:
-        from_data = _find_from_asset_and_amount(decoded_events)
-        if from_data is None:
-            return None
-
-        return SwapData(
-            from_asset=from_data[0],
-            from_amount=from_data[1],
-            to_asset=self.node_inquirer.native_token,
-            to_amount=receive_native_event.amount,
-        )
-
-    def _decode_token_to_token_swap(
-            self,
-            decoded_events: list[EvmEvent],
-    ) -> SwapData | None:
-        from_data = _find_from_asset_and_amount(decoded_events)
-        to_data = self._find_to_asset_and_amount(decoded_events)
-        if from_data is None or to_data is None:
-            return None
-
-        return SwapData(
-            from_asset=from_data[0],
-            from_amount=from_data[1],
-            to_asset=to_data[0],
-            to_amount=to_data[1],
-        )
-
     def _routers_post_decoding(
             self,
             transaction: EvmTransaction,
@@ -297,80 +173,16 @@ class Uniswapv3CommonDecoder(EvmDecoderInterface):
         """
         if transaction.to_address not in self.routers_addresses:
             return decoded_events  # work only with the known routers for now
-        if len(decoded_events) < 3:
-            # The minimum is 3 events: gas, spend, receive.
-            return decoded_events
-
-        send_native_event, receive_native_event = None, None
-        for event in decoded_events:
-            if event.asset == self.node_inquirer.native_token and event.counterparty != CPT_GAS:
-                if event.event_type == HistoryEventType.SPEND:
-                    send_native_event = event
-                else:  # Receive
-                    receive_native_event = event
-
-        # Since we check that the to_address is a known router, we can be sure that the native
-        # currency transfer event (and other events) are parts of the swap.
-        if send_native_event is not None:
-            # This is a swap from native to token
-            # receive_native_event may be or may not be None depending on whether there was a refund  # noqa: E501
-            swap_data = self._decode_native_to_token_swap(decoded_events, send_native_event, receive_native_event)  # noqa: E501
-        elif receive_native_event is not None:
-            # This is a swap from token to native
-            swap_data = self._decode_token_to_native_swap(decoded_events, receive_native_event)
-        else:
-            # Then this should be a swap from token to token
-            swap_data = self._decode_token_to_token_swap(decoded_events)
-
-        if swap_data is None:
-            log.error(f'Failed to decode a {self.node_inquirer.chain_name} uniswap swap for transaction {transaction.tx_hash!s}')  # noqa: E501
-            return decoded_events
-
-        # These should never raise any errors since `from_asset` and `to_asset` are either native
-        # currency or have already been resolved to tokens during erc20 transfers decoding.
-        from_crypto_asset = swap_data.from_asset.resolve_to_crypto_asset()
-        to_crypto_asset = swap_data.to_asset.resolve_to_crypto_asset()
-
-        gas_event = None
-        for event in decoded_events:
-            if event.counterparty == CPT_GAS:
-                gas_event = event
-                break
-
-        assert gas_event is not None, 'Gas event should always exist when interacting with a uniswap auto router'  # noqa: E501
-        gas_event.sequence_index = 0
-
-        timestamp = ts_ms_to_sec(decoded_events[0].timestamp)  # all events have same timestamp
         display_name = get_versioned_counterparty_label(CPT_UNISWAP_V3)
-        from_event = self.base.make_event(
-            tx_ref=transaction.tx_hash,
-            sequence_index=1,
-            timestamp=timestamp,
-            event_type=HistoryEventType.TRADE,
-            event_subtype=HistoryEventSubType.SPEND,
-            asset=from_crypto_asset,
-            amount=swap_data.from_amount,
-            location_label=transaction.from_address,
-            notes=f'Swap {swap_data.from_amount} {from_crypto_asset.symbol} via {display_name} auto router',  # noqa: E501
+        return decode_uniswap_v3_like_router_swap(
+            transaction=transaction,
+            decoded_events=decoded_events,
             counterparty=CPT_UNISWAP_V3,
-            address=transaction.to_address,
+            spend_notes=f'Swap {{amount}} {{symbol}} via {display_name} auto router',
+            receive_notes=f'Receive {{amount}} {{symbol}} as the result of a swap via {display_name} auto router',  # noqa: E501
+            source_counterparties={CPT_UNISWAP_V2, CPT_UNISWAP_V3},
+            retain_non_swap_events=False,
         )
-
-        to_event = self.base.make_event(
-            tx_ref=transaction.tx_hash,
-            sequence_index=2,
-            timestamp=timestamp,
-            event_type=HistoryEventType.TRADE,
-            event_subtype=HistoryEventSubType.RECEIVE,
-            asset=to_crypto_asset,
-            amount=swap_data.to_amount,
-            location_label=transaction.from_address,
-            notes=f'Receive {swap_data.to_amount} {to_crypto_asset.symbol} as the result of a swap via {display_name} auto router',  # noqa: E501
-            counterparty=CPT_UNISWAP_V3,
-            address=transaction.to_address,
-        )
-
-        return [gas_event, from_event, to_event]
 
     def _lp_post_decoding(
             self,